+++ src/main/java/com/telepathicgrunt/blame/mixin/JigsawPatternMixin.java
@@ -1,12 +1,22 @@
 package com.telepathicgrunt.blame.mixin;
 
+import com.google.common.collect.Lists;
+import com.mojang.datafixers.util.Pair;
+import com.telepathicgrunt.blame.main.JigsawPatternBlame;
 import com.telepathicgrunt.blame.main.MissingNBTBlame;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.gen.feature.jigsaw.JigsawPattern;
+import net.minecraft.world.gen.feature.jigsaw.JigsawPiece;
 import net.minecraft.world.gen.feature.template.TemplateManager;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /* @author - TelepathicGrunt
  *
@@ -17,6 +27,19 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(JigsawPattern.class)
 public class JigsawPatternMixin {
 
+	@Redirect(method = "<init>(Lnet/minecraft/util/ResourceLocation;Lnet/minecraft/util/ResourceLocation;Ljava/util/List;)V",
+			at = @At(value = "INVOKE", target = "Lcom/google/common/collect/Lists;newArrayList()Ljava/util/ArrayList;"))
+	private ArrayList<JigsawPiece> tooLargePool(ResourceLocation name, ResourceLocation fallback,
+												List<Pair<JigsawPiece, Integer>> pieceElements)
+	{
+		for(Pair<JigsawPiece, Integer> element : pieceElements){
+			if(element.getSecond() > 100000){
+				JigsawPatternBlame.printExcessiveWeight(name, element);
+			}
+		}
+
+		return new ArrayList<>();
+	}
 
 	@Inject(method = "getMaxSize(Lnet/minecraft/world/gen/feature/template/TemplateManager;)I",
 			at = @At(value = "HEAD"))