+++ src/main/java/com/telepathicgrunt/blame/mixin/StructureFeatureMixin.java
@@ -1,19 +1,11 @@
 package com.telepathicgrunt.blame.mixin;
 
-import com.google.common.collect.BiMap;
-import com.google.common.collect.HashBiMap;
-import com.telepathicgrunt.blame.main.MissingNBTBlame;
 import com.telepathicgrunt.blame.main.StructureFeatureBlame;
-import net.minecraft.structure.Structure;
-import net.minecraft.structure.StructureManager;
-import net.minecraft.util.Identifier;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.world.gen.ChunkRandom;
 import net.minecraft.world.gen.chunk.StructureConfig;
 import net.minecraft.world.gen.feature.StructureFeature;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;