+++ src/main/java/com/telepathicgrunt/blame/mixin/JigsawPatternMixin.java
@@ -28,17 +28,14 @@ import java.util.List;
 public class JigsawPatternMixin {
 
 	@Redirect(method = "<init>(Lnet/minecraft/util/ResourceLocation;Lnet/minecraft/util/ResourceLocation;Ljava/util/List;)V",
-			at = @At(value = "INVOKE", target = "Lcom/google/common/collect/Lists;newArrayList()Ljava/util/ArrayList;"))
-	private ArrayList<JigsawPiece> tooLargePool(ResourceLocation name, ResourceLocation fallback,
-												List<Pair<JigsawPiece, Integer>> pieceElements)
+			at = @At(value = "INVOKE", target = "Lcom/mojang/datafixers/util/Pair;getFirst()Ljava/lang/Object;"))
+	private <F> F tooLargePool(Pair<F, Integer> pair, ResourceLocation name, ResourceLocation fallback, List<Pair<JigsawPiece, Integer>> pieceElements)
 	{
-		for(Pair<JigsawPiece, Integer> element : pieceElements){
-			if(element.getSecond() > 100000){
-				JigsawPatternBlame.printExcessiveWeight(name, element);
-			}
+		if(pair.getSecond() > 100000){
+			JigsawPatternBlame.printExcessiveWeight(name, (Pair<JigsawPiece, Integer>) pair);
 		}
 
-		return new ArrayList<>();
+		return pair.getFirst();
 	}
 
 	@Inject(method = "getMaxSize(Lnet/minecraft/world/gen/feature/template/TemplateManager;)I",