+++ src/main/java/com/telepathicgrunt/blame/mixin/SingleJigsawPieceAccessor.java
@@ -17,5 +17,5 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 public interface SingleJigsawPieceAccessor {
 
 	@Accessor("field_236839_c_")
-	Either<ResourceLocation, Template> getTemplateRL();
+	Either<ResourceLocation, Template> blame_getTemplateRL();
 }