+++ CHANGELOG.md
@@ -10,7 +10,10 @@ Gotta blame the broken mods!
    
     JSON file on world loading:
 •  Added extra info about missing block properties from broken ConfiguredFeatures files to help modders fix them better.
-           
+    
+    Template Pools:
+• Will now detect and print out if a mod or datapack has a pool element with an insanely high weight that is eating up all the RAM.
+
            
    (V.1.8.1 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)