+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Gotta blame the broken mods!
     JSON file on world loading:
 •  Added extra info about missing block properties from broken ConfiguredFeatures files to help modders fix them better.
       
+    Template Pools:
+• Will now detect and print out if a mod or datapack has a pool element with an insanely high weight that is eating up all the RAM.
+
       
    (V.2.3.0 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)